requirements.txt
scikit-image~=0.17.2
numpy~=1.19.2
matplotlib~=3.3.4
pillow~=8.1.0
tqdm~=4.56.0
scikit-learn~=0.24.1